inplace, achieve perfect speedup

+++ src/metrics.hpp
@@ -4,17 +4,17 @@
 #include <numeric>
 #include <algorithm>
 
-using Matrix = std::vector<std::vector<float>>;
+using Matrix = std::vector<std::vector<double>>;
 
 // Mean Squared Error (MSE)
-inline float compute_mse(const Matrix& img1, const Matrix& img2) {
+inline double compute_mse(const Matrix& img1, const Matrix& img2) {
     int rows = img1.size();
     int cols = img1[0].size();
-    float mse = 0.0f;
+    double mse = 0.0f;
 
     for (int i = 0; i < rows; ++i)
         for (int j = 0; j < cols; ++j) {
-            float diff = img1[i][j] - img2[i][j];
+            double diff = img1[i][j] - img2[i][j];
             mse += diff * diff;
         }
 
@@ -22,13 +22,13 @@ inline float compute_mse(const Matrix& img1, const Matrix& img2) {
 }
 
 // Structural Similarity Index Measure (SSIM) simplified
-inline float compute_ssim(const Matrix& img1, const Matrix& img2) {
-    constexpr float C1 = 6.5025f, C2 = 58.5225f;
+inline double compute_ssim(const Matrix& img1, const Matrix& img2) {
+    constexpr double C1 = 6.5025f, C2 = 58.5225f;
 
     int rows = img1.size();
     int cols = img1[0].size();
 
-    float mean_x = 0, mean_y = 0, sigma_x = 0, sigma_y = 0, sigma_xy = 0;
+    double mean_x = 0, mean_y = 0, sigma_x = 0, sigma_y = 0, sigma_xy = 0;
 
     // Compute means
     for (int i = 0; i < rows; ++i)
@@ -42,8 +42,8 @@ inline float compute_ssim(const Matrix& img1, const Matrix& img2) {
     // Compute variances and covariance
     for (int i = 0; i < rows; ++i)
         for (int j = 0; j < cols; ++j) {
-            float dx = img1[i][j] - mean_x;
-            float dy = img2[i][j] - mean_y;
+            double dx = img1[i][j] - mean_x;
+            double dy = img2[i][j] - mean_y;
             sigma_x += dx * dx;
             sigma_y += dy * dy;
             sigma_xy += dx * dy;
@@ -53,8 +53,8 @@ inline float compute_ssim(const Matrix& img1, const Matrix& img2) {
     sigma_xy /= (rows * cols - 1);
 
     // Compute SSIM
-    float numerator = (2 * mean_x * mean_y + C1) * (2 * sigma_xy + C2);
-    float denominator = (mean_x * mean_x + mean_y * mean_y + C1) * (sigma_x + sigma_y + C2);
+    double numerator = (2 * mean_x * mean_y + C1) * (2 * sigma_xy + C2);
+    double denominator = (mean_x * mean_x + mean_y * mean_y + C1) * (sigma_x + sigma_y + C2);
 
     return numerator / denominator;
 }
